Revert "Add world interaction recipes for straw cutting"
This reverts commit 5713b4e1428988e49e502115afda570822bcb1e4.
World interaction is only used for right-click interactions. Also it's not like there's a recipe for obsidian that says it needs a diamond pick. Knives cutting straw is in the quest book, which should be enough info for players.

A better alternative for the future is making a GT tool behavior for knives called "Cutter", which would display in the knife tooltip like "Cutter: Cuts Straw".