Re-enable disabled blocks for crafting/decoration
Furnaces, blast furnaces, and smokers are common blocks to decorate with, and they also are commonly found in structures. Instead of removing them and limiting builders, a tooltip on those blocks tells the player that the block has no functionality and is only used for crafting or decoration.

Campfires aren't a good replacement for furnaces in many recipes (especially for the furnace minecart recipe), so I have reverted that change.

I also moved furnace, blast furnace, and smoker disabling to GBTHCore to fix an issue where blocks couldn't be placed on them with shift-click.